Add templated metrics

+++ src/LinqAn.Google/Metrics/Goal17Completions.cs
@@ -6,7 +6,7 @@ namespace LinqAn.Google.Metrics
 	/// 	The total number of completions for the requested goal number.
 	/// </summary>
 	[Description("The total number of completions for the requested goal number.")]
-	public class Goal17Completions: Metric<float>
+	public class Goal17Completions: Metric<int>
 	{
 		/// <summary>
 		/// 	Instantiates a <seealso cref="Goal17Completions" />.
